requirements.txt
Flask>=3.1.1
Flask-CORS>=6.0.1
python-docx>=1.2.0
lxml>=3.1.0
spacy>=3.8.7
textstat>=0.7.8
requests>=2.32.4
python-dotenv>=1.1.1
Werkzeug>=3.0.1 